Fix #2363
