+++ facet-json/tests/integration/mod.rs
@@ -1,7 +1,6 @@
 #[path = "backend_historical.rs"]
 pub(crate) mod json_backend;
 
-mod field_proxy_untagged_bool_repro;
 mod flatten_defaults;
 mod flatten_in_externally_tagged_enum;
 mod format_specific_proxy;
@@ -24,6 +23,7 @@ mod issue_2004;
 mod issue_2007;
 mod issue_2010;
 mod issue_2059;
+mod issue_2363_bool_string_number_field_proxy;
 mod list_deferred_processing;
 mod metadata_container_flatten_map;
 mod mixed_tagged_untagged;
